LayoutTests/editing/selection/extend-selection-backward-at-beginning-of-line-by-sentence-granularity-expected.txt

This tests extending the selection backwards at the beginning of a line by sentence granularity. The first sentence should be selected.
| "\n"
| <div>
|   "<#selection-focus>This is the first sentence"
| "\n"
| <div>
|   id="target"
|   <#selection-anchor>
|   "Here is another one."
| "\n"
